Database article and publication tables updated
The auto increment IDs of the publication table are now foreign keys in
the article table.
For this: needed a second query that takes the auto-increment id of the
publication table and in case the record for pid is the same as the one
that we want to add it won't do anything if it isn't then the record
will be inserted.