Tagger cleaning
The tagger and lemmatizing was tagging even words that contained numbers
or characters. Fixed this before adding the noun/verb to their lists by
using a regex expression that takes only words that contain letters and
dashes and that contain more than one letter.

Started the labelling helper as a side application.