Labelling Helper created
The Labelling Helper takes a random number and checks the article with
that specific ID and shows text and title if the corresponding
article_id is not recorded in the articledomains table.
Takes the user input as a number from 1 to 5 (the domain ids) and
inserts it to the database in the article domain table with the random
id of the article.
This is done recursively.

Also changed a bug with the page label in the main project (changed from
int to String the database)