Creating ML tables
The training and test data tables are created. This was done using a
StringBuilder to create the SQL statement. This is adding ass columns
the verbs that appear more than 5 times and are in the articles that
have been labelled. This was done using a loop that goes through an
array list of verbs. The array list is extracted from the database using
a SELECT SQL statement.