Creating ML table refactoring
Changed the way the ML tables are created to have less duplicated code
and to reduce the number of times the SELECT query is called. Added the
domain columns
Also part of the data was inserted in the table by calling the article
id that was labelled and getting the attributes needed from the article
table.